author	David Woodhouse <David.Woodhouse@intel.com>	Wed, 12 Feb 2014 16:56:06 +0100 (15:56 +0000)
The SOCKET type is a pointer, although in practice what's returned really does look like a file descriptor. It's a low-valued integer such as 0x23 under Wine, 0x54 under Windows 7 in my testing.

The INVALID_SOCKET error return from socket() is defined as (SOCKET)(~0), or 0xFFFFFFFFFFFFFFFF on Win64. Thus we get errors when assigning it to an int in the socks[] array. So use -1 here instead.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>